Bump Version number of 0.4.0.0
This is being done for compatibility purposes with the master build of
Euterpea, but I am not pushing this to Hackage just yet.
In general, it is probably a good strategy to push the major version
number any time this repo differs in a major way from the latest Hackage
build.